Path Planning
In this project we used the the 3 link robot and the setup for controling the robot from the Matlab GUI to help us create a path planning robot, we used the Dstar, Astar, and PRM algorithams.
In thei we had push buttons that would throw obsticles and the robot would have to path plan accordingly. The video for this project is here:https://www.youtube.com/watch?v=md7zKonyrdE